+++ customers/my_script.py
@@ -1,0 +1,5 @@
+"""Used to for app manual testing"""
+
+if __name__ == '__main__':
+    from customers.core.config import DevelopmentConfig, TestingConfig
+    import ipdb; ipdb.set_trace()